requirements.txt
streamlit
fastapi
uvicorn
pydantic
requests
ollama
twilio
langgraph
langchain
langchain_groq
python-dotenv
langchain-groq